Software Engineering/Full Stack UNCA Schedule Planner/backend/pyproject.toml
[tool.poetry]
name = "Final Project"
version = "0.0.0"
description = "App for CSCI338"
authors = ["Sarah Van Wart <svanwart@unca.edu>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.101.1"
uvicorn = "^0.23.2"
aiohttp = "3.8.2"
aiofiles = "^23.2.1"
httpx = "^0.25.0"
pytest = "^7.4.2"
asyncpg = "^0.28.0"
SQLAlchemy = "^2.0.0"
psycopg2 = "^2.9.9"
Faker = "^19.12.0"
python-dotenv = "^1.0.0"
requests = "^2.32.3"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
flake8 = "^7.1.1"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"